RA-1174 - :hammer: Moving get vacancy orchestrators for apprenticeships and traineeships under the correct test feature unit testing folder.
