+++ src/Esfa.Vacancy.Register.UnitTests/GetApprenticeshipVacancy/Api/Orchestrators/GivenAGetApprenticeshipVacancyOrchestrator/WhenGettingLiveApprenticeship.cs
@@ -1,7 +1,8 @@
-﻿using System.Threading;
+using System.Threading;
 using System.Threading.Tasks;
+using Esfa.Vacancy.Api.Types;
 using Esfa.Vacancy.Register.Api.Orchestrators;
-using Esfa.Vacancy.Register.Application.Queries.GetTraineeshipVacancy;
+using Esfa.Vacancy.Register.Application.Queries.GetApprenticeshipVacancy;
 using Esfa.Vacancy.Register.Infrastructure.Settings;
 using FluentAssertions;
 using MediatR;
@@ -9,33 +10,34 @@ using Moq;
 using NUnit.Framework;
 using Ploeh.AutoFixture;
 
-namespace Esfa.Vacancy.Register.UnitTests.Api.Orchestrators.TraineeshipVacancyOrchestratorTests
+namespace Esfa.Vacancy.Register.UnitTests.GetApprenticeshipVacancy.Api.Orchestrators.GivenAGetApprenticeshipVacancyOrchestrator
 {
     [TestFixture]
     [Parallelizable(ParallelScope.Fixtures)]
-    public class WhenGettingLiveTraineeship
+    public class WhenGettingLiveApprenticeship
     {
         private const int VacancyReference = 1234;
         private const int LiveVacancyStatusId = 2;
         private Mock<IMediator> _mockMediator;
-        private Mock<IProvideSettings> _mockProvideSettings;
-        private GetTraineeshipVacancyOrchestrator _sut;
-        
+        private Mock<IProvideSettings> _provideSettings;
+        private GetApprenticeshipVacancyOrchestrator _sut;
+
         [SetUp]
         public void SetUp()
         {
             _mockMediator = new Mock<IMediator>();
-            _mockProvideSettings = new Mock<IProvideSettings>();
-            _sut = new GetTraineeshipVacancyOrchestrator(_mockMediator.Object, _mockProvideSettings.Object);
+            _provideSettings = new Mock<IProvideSettings>();
+
+            _sut = new GetApprenticeshipVacancyOrchestrator(_mockMediator.Object, _provideSettings.Object);
         }
 
         [Test]
-        public async Task WithNonAnonymousEmployer_ShouldNotReplaceEmployerNameAndDescription()
+        public async Task GetLiveNonAnonymousEmployerVacancy_ShouldNotReplaceEmployerNameAndDescription()
         {
-            _mockMediator.Setup(m => m.Send(It.IsAny<GetTraineeshipVacancyRequest>(), CancellationToken.None))
-                .ReturnsAsync(new GetTraineeshipVacancyResponse
+            _mockMediator.Setup(m => m.Send(It.IsAny<GetApprenticeshipVacancyRequest>(), CancellationToken.None))
+                .ReturnsAsync(new GetApprenticeshipVacancyResponse
                 {
-                    TraineeshipVacancy = new Fixture().Build<Domain.Entities.TraineeshipVacancy>()
+                    ApprenticeshipVacancy = new Fixture().Build<Domain.Entities.ApprenticeshipVacancy>()
                                             .With(v => v.VacancyReferenceNumber, VacancyReference)
                                             .With(v => v.VacancyStatusId, LiveVacancyStatusId)
                                             .With(v => v.EmployerName, "ABC Ltd")
@@ -47,7 +49,7 @@ namespace Esfa.Vacancy.Register.UnitTests.Api.Orchestrators.TraineeshipVacancyOr
                                             .Create()
                 });
 
-            var result = await _sut.GetTraineeshipVacancyDetailsAsync(VacancyReference);
+            var result = await _sut.GetApprenticeshipVacancyDetailsAsync(VacancyReference);
 
             result.VacancyReference.Should().Be(VacancyReference);
             result.EmployerName.Should().Be("ABC Ltd");
@@ -66,12 +68,12 @@ namespace Esfa.Vacancy.Register.UnitTests.Api.Orchestrators.TraineeshipVacancyOr
         }
 
         [Test]
-        public async Task WithAnonymousEmployer_ShouldReplaceEmployerNameAndDescription()
+        public async Task GetLiveAnonymousEmployerVacancy_ShouldReplaceEmployerNameAndDescription()
         {
-            _mockMediator.Setup(m => m.Send(It.IsAny<GetTraineeshipVacancyRequest>(), CancellationToken.None))
-                .ReturnsAsync(new GetTraineeshipVacancyResponse
+            _mockMediator.Setup(m => m.Send(It.IsAny<GetApprenticeshipVacancyRequest>(), CancellationToken.None))
+                .ReturnsAsync(new GetApprenticeshipVacancyResponse
                 {
-                    TraineeshipVacancy = new Fixture().Build<Domain.Entities.TraineeshipVacancy>()
+                    ApprenticeshipVacancy = new Fixture().Build<Domain.Entities.ApprenticeshipVacancy>()
                                             .With(v => v.VacancyReferenceNumber, VacancyReference)
                                             .With(v => v.VacancyStatusId, LiveVacancyStatusId)
                                             .With(v => v.EmployerName, "Her Majesties Secret Service")
@@ -82,7 +84,7 @@ namespace Esfa.Vacancy.Register.UnitTests.Api.Orchestrators.TraineeshipVacancyOr
                                             .Create()
                 });
 
-            var result = await _sut.GetTraineeshipVacancyDetailsAsync(VacancyReference);
+            var result = await _sut.GetApprenticeshipVacancyDetailsAsync(VacancyReference);
 
             result.VacancyReference.Should().Be(VacancyReference);
             result.EmployerName.Should().Be("ABC Ltd");
@@ -100,30 +102,56 @@ namespace Esfa.Vacancy.Register.UnitTests.Api.Orchestrators.TraineeshipVacancyOr
         }
 
         [Test]
-        public async Task ShouldPopulateTraineeshipVacancyUrl()
+        public async Task ShouldPopulateVacancyUrl()
         {
             //Arrange
-            var baseUrl = "https://findapprentice.com/traineeship/reference";
+            var baseUrl = "https://findapprentice.com/apprenticeship/reference";
 
-            _mockProvideSettings.Setup(p => p.GetSetting(ApplicationSettingKeyConstants.LiveTraineeshipVacancyBaseUrlKey))
-                                .Returns(baseUrl);
-
-            var response = new GetTraineeshipVacancyResponse
+            _provideSettings.Setup(p => p.GetSetting(ApplicationSettingKeyConstants.LiveApprenticeshipVacancyBaseUrlKey)).Returns(baseUrl);
+            
+            var response = new GetApprenticeshipVacancyResponse()
             {
-                TraineeshipVacancy = new Fixture().Build<Domain.Entities.TraineeshipVacancy>()
-                                                    .With(v => v.VacancyReferenceNumber, VacancyReference)
-                                                    .Create()
+                ApprenticeshipVacancy = new Fixture().Build<Domain.Entities.ApprenticeshipVacancy>()
+                                            .With(v => v.VacancyReferenceNumber, VacancyReference)
+                                            .Create()
             };
 
             _mockMediator
-                .Setup(m => m.Send(It.IsAny<GetTraineeshipVacancyRequest>(), It.IsAny<CancellationToken>()))
+                .Setup(m => m.Send(It.IsAny<GetApprenticeshipVacancyRequest>(), It.IsAny<CancellationToken>()))
                 .ReturnsAsync(response);
 
+            var sut = new GetApprenticeshipVacancyOrchestrator(_mockMediator.Object, _provideSettings.Object);
             //Act
-            var vacancy = await _sut.GetTraineeshipVacancyDetailsAsync(VacancyReference);
+            var vacancy = await sut.GetApprenticeshipVacancyDetailsAsync(12345);
 
             //Assert
             Assert.AreEqual($"{baseUrl}/{VacancyReference}", vacancy.VacancyUrl);
+        }
+
+        [TestCase(1, WageUnit.Unspecified)]
+        [TestCase(2, WageUnit.Weekly)]
+        [TestCase(3, WageUnit.Monthly)]
+        [TestCase(4, WageUnit.Annually)]
+        [TestCase(null, WageUnit.Unspecified)]
+        public async Task ShouldMapWageUnitCorrectly(int? wageUnitId, WageUnit? wageUnitType)
+        {
+            //Arrange
+            var response = new GetApprenticeshipVacancyResponse()
+            {
+                ApprenticeshipVacancy = new Fixture().Build<Domain.Entities.ApprenticeshipVacancy>()
+                                        .With(v => v.WageUnitId, wageUnitId)
+                                        .Create()
+            };
+
+            _mockMediator
+                .Setup(m => m.Send(It.IsAny<GetApprenticeshipVacancyRequest>(), It.IsAny<CancellationToken>()))
+                .ReturnsAsync(response);
+
+            var sut = new GetApprenticeshipVacancyOrchestrator(_mockMediator.Object, _provideSettings.Object);
+            //Act
+            var vacancy = await sut.GetApprenticeshipVacancyDetailsAsync(VacancyReference);
+            //Assert
+            vacancy.WageUnit.ShouldBeEquivalentTo(wageUnitType);
         }
     }
 }
